Move pyramid exception classes into a dedicated exceptions sub-package
NoMatchingResolutionException, NotAMultiscaleImageException and
NotASingleScaleImageException are domain-level error types that the
pyramid package and its callers will all share once additional backend
implementations land. Grouping them under sc.fiji.ome.zarr.pyramid.exceptions
keeps the top-level pyramid package focused on the user-facing API.
